requirements.txt
anyio==3.6.1
asgiref==3.5.2
astroid==2.11.7
backports.cached-property==1.0.2
black==22.3.0
click==8.1.3
coverage==6.4.3
decorator==5.1.1
dill==0.3.5.1
fastapi==0.78.0
graphql-core==3.2.1
h11==0.13.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
powerline-status==2.7
pydantic==1.9.2
Pygments==2.12.0
pylint==2.14.2
pytemperature==1.1
python-dateutil==2.8.2
python-multipart==0.0.5
six==1.15.0
sniffio==1.2.0
starlette==0.19.1
strawberry-graphql==0.114.2
tomli==2.0.1
tomlkit==0.11.3
typing_extensions==4.3.0
uvicorn==0.17.6
validators==0.20.0
wrapt==1.14.1
